// apps/web/src/api/model/apiHydratedBilling.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Allawee Infrastructure Backend APIs
 * The Allawee Infrastructure APIs description
 * OpenAPI spec version: 0.0.1
 */
import type { ApiHydratedBillingStatus } from './apiHydratedBillingStatus';
import type { AWSObjectURL } from './aWSObjectURL';
import type { BillingItem } from './billingItem';

export interface ApiHydratedBilling {
  /** Invoice number */
  invoiceNo: string;
  /** Billing period (YYYY-MM format) */
  period: string;
  /** List (pre-discount) amount */
  listAmount: number;
  /** Billed amount */
  billedAmount: number;
  /** Invoice due date */
  dueDate: string;
  /** Billing line items */
  items: BillingItem[];
  /** Billing currency */
  currency: string;
  /** Billing status */
  status: ApiHydratedBillingStatus;
  /** Invoice PDF */
  invoicePdf?: AWSObjectURL;
  /** Document id */
  id: string;
  /** Document created at */
  createdAt: string;
  /** Document updated at */
  updatedAt: string;
}